Make CRPS comparison relative. Add to docker image.
