+++ Makefile
@@ -105,7 +105,7 @@ docs-serve:
 # Example DOCKER_REPO?=nvcr.io/dycvht5ows21
 E2S_RELEASE_TAG?=0.15.0
 E2S_IMAGE_NAME=$(DOCKER_REPO)/earth2studio-scicomp
-E2S_IMAGE_TAG=v$(E2S_RELEASE_TAG).20260512.2
+E2S_IMAGE_TAG=v$(E2S_RELEASE_TAG).20260513.0
 container-service:
 	@test -n "$(DOCKER_REPO)" || (echo "DOCKER_REPO is not set!" && exit 1)
 	DOCKER_BUILDKIT=1 docker build -t $(E2S_IMAGE_NAME):$(E2S_IMAGE_TAG) -f serve/Dockerfile .
